1、TASK_831(PC端报告图片删除时增加是否确认删除提示)

+++ src/pages/index/packages/diseaseModule/CRFConfig/display/displayFileUpload.vue
@@ -192,16 +192,22 @@
           })
         },
         handleRemove(file,fileList) {
-          console.log(file,fileList);
-          this.deleteFileId(file.fileId);
-          let copyFileList  = this.fileList;
-          copyFileList.forEach((item,index,array)=>{
-            if(item.fileId == file.fileId) {
-              array.splice(index,1)
-            }
+          this.$confirm('此操作将删除该信息, 是否继续?', '提示', {
+            confirmButtonText: '确定',
+            cancelButtonText: '取消',
+            type: 'warning'
+          }).then(() => {
+            this.deleteFileId(file.fileId);
+            let copyFileList  = this.fileList;
+            copyFileList.forEach((item,index,array)=>{
+              if(item.fileId == file.fileId) {
+                array.splice(index,1)
+              }
+            });
+            this.fileList = copyFileList;
+            this.report.value2 = JSON.stringify(this.fileList);
+          }).catch(() => {
           });
-          this.fileList = copyFileList;
-          this.report.value2 = JSON.stringify(this.fileList);
         },
         closeViewer() {
           this.dialogVisible = false;
